Overhaul dev package builds

+++ devtools/osx-build.sh
@@ -9,16 +9,18 @@ curl -s -O https://repo.continuum.io/miniconda/Miniconda3-latest-MacOSX-x86_64.s
 bash Miniconda3-latest-MacOSX-x86_64.sh -b -p $HOME/anaconda;
 export PATH=$HOME/anaconda/bin:$PATH;
 conda config --add channels omnia;
-conda install -yq conda-build jinja2 anaconda-client;
+conda install -yq conda-build=1.21.3 jinja2 anaconda-client;
 
-# Install OpenMM dependencies that can't be installed through
-# conda package manager (doxygen + CUDA)
-brew install -y --quiet doxygen
-curl -O -s http://developer.download.nvidia.com/compute/cuda/7.5/Prod/network_installers/mac/x86_64/cuda_mac_installer_tk.tar.gz
-curl -O -s http://developer.download.nvidia.com/compute/cuda/7.5/Prod/network_installers/mac/x86_64/cuda_mac_installer_drv.tar.gz
-sudo tar -zxf cuda_mac_installer_tk.tar.gz -C /;
-sudo tar -zxf cuda_mac_installer_drv.tar.gz -C /;
-rm -f cuda_mac_installer_tk.tar.gz cuda_mac_installer_drv.tar.gz
+if ! ./conda-build-all --dry-run -- openmm* ; then
+    # Install OpenMM dependencies that can't be installed through
+    # conda package manager (doxygen + CUDA)
+    brew install -y --quiet doxygen
+    curl -O -s http://developer.download.nvidia.com/compute/cuda/7.5/Prod/network_installers/mac/x86_64/cuda_mac_installer_tk.tar.gz
+    curl -O -s http://developer.download.nvidia.com/compute/cuda/7.5/Prod/network_installers/mac/x86_64/cuda_mac_installer_drv.tar.gz
+    sudo tar -zxf cuda_mac_installer_tk.tar.gz -C /;
+    sudo tar -zxf cuda_mac_installer_drv.tar.gz -C /;
+    rm -f cuda_mac_installer_tk.tar.gz cuda_mac_installer_drv.tar.gz
+fi;
 
 # Install latex.
 brew tap -y --quiet Caskroom/cask;
@@ -28,4 +30,8 @@ sudo tlmgr update --self
 sudo tlmgr install titlesec framed threeparttable wrapfig multirow collection-fontsrecommended hyphenat xstring
 
 # Build packages
-./conda-build-all $CONDA_BUILD_ALL_FLAGS *;
+if [[ "${TRAVIS_PULL_REQUEST}" == "false" ]]; then
+    ./conda-build-all -vvv $UPLOAD  -- * || true;
+else
+    ./conda-build-all -vvv $UPLOAD -- *;
+fi;
